ai_service/requirements.txt
# Python 3.13 Compatible Requirements
# For Python 3.13, we need to use versions that have pre-built wheels available
# Install with: pip install -r requirements-py313.txt

# Core web framework
fastapi>=0.115.0

# ASGI server
uvicorn[standard]>=0.32.0

# Data validation - use latest with Python 3.13 wheels
pydantic>=2.10.0

# LangChain - use latest versions with Python 3.13 support
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# LangGraph - use latest
langgraph>=0.2.0

# Anthropic API - use latest with Python 3.13 wheels
anthropic>=0.40.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0

# Windows console colors
colorama>=0.4.6